Projekt/requirements.txt
Flask>=3.0
mysql-connector-python>=8.4
python-dotenv>=1.0
reportlab>=4.0
Flask-Mail>=0.9
qrcode[pil]
Pillow